Added JWT auth to product API's and tested locally

+++ server/routes/productRoute.js
@@ -44,22 +44,20 @@ router.get('/:id', auth, getProduct, (req, res) => {
   
 // PATCH route to update a product's details
 // USES :: get product Middleware 
-router.patch('/:id', auth, async (req, res) => {
+router.patch('/:id', auth, getProduct, async (req, res) => {
         
     try {
-        const product = await Product.findOne({ _id: req.params.id, user: req.user._id });
-        // console.log(' id: ', req.params.id, 'user: ', req.user._id);
+        const product = await Product.findOne({ _id: req.params.id , user: req.user._id});         
         if (!product) {
           return res.status(404).json({ message: 'Product not found or user not authorized' });
         }
         
         // Dynamically update fields provided in the request
         Object.entries(req.body).forEach(([key, value]) => {
-          product[key] = value;
-      });
-        console.log('Product ', updatedProduct);
-        const updatedProduct = await res.product.save();
-        res.json(updatedProduct);
+          res.product[key] = value;
+        });      
+      const updatedProduct = await res.product.save();
+      res.json(updatedProduct);
     } catch (err) {
         res.status(400).json({ message: err.message });
         // console.log('error here', err.message);
@@ -68,14 +66,17 @@ router.patch('/:id', auth, async (req, res) => {
 
 
 // DELETE route to delete a product
-router.delete('/:id', getProduct, async (req, res) => {
+router.delete('/:id',auth, getProduct, async (req, res) => {
     try {
-        const deletedProduct = await Product.findByIdAndDelete(req.params.id);
+        const deletedProduct = await Product.findOne({ _id: req.params.id });
         if (!deletedProduct) {
-            return res.status(404).json({ message: 'Cannot find product' });
+          return res.status(404).json({ message: 'Cannot find product' });
         }
-
-    res.json({ message: 'Deleted Product' });
+        if (deletedProduct.user.toString() !== req.user.id) {
+          return res.status(403).json({ message: 'User not authorized to delete this product' });
+        }
+        await Product.findByIdAndDelete(req.params.id);
+    res.json({ message: 'Deleted Product ', product: deletedProduct.name });
     } catch (err) {
     res.status(500).json({ message: err.message });
     }
@@ -90,11 +91,11 @@ async function getProduct(req, res, next) {
     if (product == null) {
         return res.status(404).json({ message: 'Cannot find product' });
     }
+    res.product = product;
+    next();
     } catch (err) {
         return res.status(500).json({ message: err.message });
     }
-    res.product = product;
-    next();
 }
 
 module.exports = router;
